+++ application/DevRetrofitUse/src/main/java/afkt/retrofit/use/simple/SimpleFragment.kt
@@ -24,10 +24,10 @@ class SimpleFragment : BaseFragment<FragmentSimpleBinding, SimpleViewModel>(
     override fun initObserve() {
         super.initObserve()
         // 监听电影详情信息变更
-        viewModel.movieDetailOB.state.observe(this) { value ->
+        viewModel.movieDetailOB.value.observe(this) { value ->
             // 打印日志
             ResponseHelper.log(
-                "movieDetail StateLiveData",
+                "movieDetail ValueLiveData",
                 "设置 ObservableField<MovieDetailBean> 值：${value}"
             )
         }