+++ application/DevUtilsApp/src/main/java/afkt/project/features/other_function/ActivityResultAPIFragment.kt
@@ -22,7 +22,7 @@ import dev.engine.extensions.log.log_dTag
 import dev.engine.extensions.permission.permission_request
 import dev.engine.extensions.toast.toast_showShort
 import dev.engine.permission.IPermissionEngine
-import dev.simple.core.livedata.StateLiveData
+import dev.simple.core.livedata.ValueLiveData
 import dev.simple.extensions.hi.hiif.hiIfNotNull
 import dev.simple.extensions.toSource
 import dev.utils.app.ActivityUtils
@@ -47,7 +47,7 @@ class ActivityResultAPIFragment :
 class ActivityResultAPIViewModel : AppViewModel() {
 
     // 拍照图片 Uri
-    val imageUri = StateLiveData(DevSource.create(""))
+    val imageUri = ValueLiveData(DevSource.create(""))
 
     // 如果使用 by lazy 要确保在 Lifecycle STARTED 状态前初始化
     private var mAssist: ActivityResultAssist<Uri, Boolean>? = null
@@ -60,7 +60,7 @@ class ActivityResultAPIViewModel : AppViewModel() {
             caller, ActivityResultContracts.TakePicture()
         ) {
             if (it) {
-                imageUri.setState(mAssist?.inputValue?.toSource())
+                imageUri.setValue(mAssist?.inputValue?.toSource())
             } else {
                 toast_showShort(text = "非成功操作")
             }