+++ lib/DevSimple/src/main/java/dev/simple/core/livedata/StateIntLiveData.kt
@@ -133,7 +133,7 @@ object StateInt {
  */
 open class StateIntLiveData(
     _value: Int = StateInt.NORMAL
-) : StateLiveData<Int>(_value) {
+) : ValueLiveData<Int>(_value) {
 
     /**
      * 是否 XXX 状态
@@ -148,7 +148,7 @@ open class StateIntLiveData(
      * @return `true` success, `false` fail
      */
     open fun setXXX(state: Int): Boolean {
-        return setState(state)
+        return setValue(state)
     }
 
     /**
@@ -156,7 +156,7 @@ open class StateIntLiveData(
      * @return `true` success, `false` fail
      */
     open fun postXXX(state: Int): Boolean {
-        return postState(state)
+        return postValue(state)
     }
 
     /**
@@ -164,248 +164,248 @@ open class StateIntLiveData(
      * @return `true` success, `false` fail
      */
     open fun smartUpdateXXX(state: Int): Boolean {
-        return smartUpdateState(state)
+        return smartUpdateValue(state)
     }
 
     // =
 
     // NORMAL
     open fun isNORMALState(): Boolean = isEqual(StateInt.NORMAL)
-    open fun setNORMAL(): Boolean = setState(StateInt.NORMAL)
-    open fun postNORMAL(): Boolean = postState(StateInt.NORMAL)
-    open fun smartUpdateNORMAL(): Boolean = smartUpdateState(StateInt.NORMAL)
+    open fun setNORMAL(): Boolean = setValue(StateInt.NORMAL)
+    open fun postNORMAL(): Boolean = postValue(StateInt.NORMAL)
+    open fun smartUpdateNORMAL(): Boolean = smartUpdateValue(StateInt.NORMAL)
 
     // ING
     open fun isINGState(): Boolean = isEqual(StateInt.ING)
-    open fun setING(): Boolean = setState(StateInt.ING)
-    open fun postING(): Boolean = postState(StateInt.ING)
-    open fun smartUpdateING(): Boolean = smartUpdateState(StateInt.ING)
+    open fun setING(): Boolean = setValue(StateInt.ING)
+    open fun postING(): Boolean = postValue(StateInt.ING)
+    open fun smartUpdateING(): Boolean = smartUpdateValue(StateInt.ING)
 
     // SUCCESS
     open fun isSUCCESSState(): Boolean = isEqual(StateInt.SUCCESS)
-    open fun setSUCCESS(): Boolean = setState(StateInt.SUCCESS)
-    open fun postSUCCESS(): Boolean = postState(StateInt.SUCCESS)
-    open fun smartUpdateSUCCESS(): Boolean = smartUpdateState(StateInt.SUCCESS)
+    open fun setSUCCESS(): Boolean = setValue(StateInt.SUCCESS)
+    open fun postSUCCESS(): Boolean = postValue(StateInt.SUCCESS)
+    open fun smartUpdateSUCCESS(): Boolean = smartUpdateValue(StateInt.SUCCESS)
 
     // FAIL
     open fun isFAILState(): Boolean = isEqual(StateInt.FAIL)
-    open fun setFAIL(): Boolean = setState(StateInt.FAIL)
-    open fun postFAIL(): Boolean = postState(StateInt.FAIL)
-    open fun smartUpdateFAIL(): Boolean = smartUpdateState(StateInt.FAIL)
+    open fun setFAIL(): Boolean = setValue(StateInt.FAIL)
+    open fun postFAIL(): Boolean = postValue(StateInt.FAIL)
+    open fun smartUpdateFAIL(): Boolean = smartUpdateValue(StateInt.FAIL)
 
     // ERROR
     open fun isERRORState(): Boolean = isEqual(StateInt.ERROR)
-    open fun setERROR(): Boolean = setState(StateInt.ERROR)
-    open fun postERROR(): Boolean = postState(StateInt.ERROR)
-    open fun smartUpdateERROR(): Boolean = smartUpdateState(StateInt.ERROR)
+    open fun setERROR(): Boolean = setValue(StateInt.ERROR)
+    open fun postERROR(): Boolean = postValue(StateInt.ERROR)
+    open fun smartUpdateERROR(): Boolean = smartUpdateValue(StateInt.ERROR)
 
     // START
     open fun isSTARTState(): Boolean = isEqual(StateInt.START)
-    open fun setSTART(): Boolean = setState(StateInt.START)
-    open fun postSTART(): Boolean = postState(StateInt.START)
-    open fun smartUpdateSTART(): Boolean = smartUpdateState(StateInt.START)
+    open fun setSTART(): Boolean = setValue(StateInt.START)
+    open fun postSTART(): Boolean = postValue(StateInt.START)
+    open fun smartUpdateSTART(): Boolean = smartUpdateValue(StateInt.START)
 
     // RESTART
     open fun isRESTARTState(): Boolean = isEqual(StateInt.RESTART)
-    open fun setRESTART(): Boolean = setState(StateInt.RESTART)
-    open fun postRESTART(): Boolean = postState(StateInt.RESTART)
-    open fun smartUpdateRESTART(): Boolean = smartUpdateState(StateInt.RESTART)
+    open fun setRESTART(): Boolean = setValue(StateInt.RESTART)
+    open fun postRESTART(): Boolean = postValue(StateInt.RESTART)
+    open fun smartUpdateRESTART(): Boolean = smartUpdateValue(StateInt.RESTART)
 
     // END
     open fun isENDState(): Boolean = isEqual(StateInt.END)
-    open fun setEND(): Boolean = setState(StateInt.END)
-    open fun postEND(): Boolean = postState(StateInt.END)
-    open fun smartUpdateEND(): Boolean = smartUpdateState(StateInt.END)
+    open fun setEND(): Boolean = setValue(StateInt.END)
+    open fun postEND(): Boolean = postValue(StateInt.END)
+    open fun smartUpdateEND(): Boolean = smartUpdateValue(StateInt.END)
 
     // PAUSE
     open fun isPAUSEState(): Boolean = isEqual(StateInt.PAUSE)
-    open fun setPAUSE(): Boolean = setState(StateInt.PAUSE)
-    open fun postPAUSE(): Boolean = postState(StateInt.PAUSE)
-    open fun smartUpdatePAUSE(): Boolean = smartUpdateState(StateInt.PAUSE)
+    open fun setPAUSE(): Boolean = setValue(StateInt.PAUSE)
+    open fun postPAUSE(): Boolean = postValue(StateInt.PAUSE)
+    open fun smartUpdatePAUSE(): Boolean = smartUpdateValue(StateInt.PAUSE)
 
     // RESUME
     open fun isRESUMEState(): Boolean = isEqual(StateInt.RESUME)
-    open fun setRESUME(): Boolean = setState(StateInt.RESUME)
-    open fun postRESUME(): Boolean = postState(StateInt.RESUME)
-    open fun smartUpdateRESUME(): Boolean = smartUpdateState(StateInt.RESUME)
+    open fun setRESUME(): Boolean = setValue(StateInt.RESUME)
+    open fun postRESUME(): Boolean = postValue(StateInt.RESUME)
+    open fun smartUpdateRESUME(): Boolean = smartUpdateValue(StateInt.RESUME)
 
     // STOP
     open fun isSTOPState(): Boolean = isEqual(StateInt.STOP)
-    open fun setSTOP(): Boolean = setState(StateInt.STOP)
-    open fun postSTOP(): Boolean = postState(StateInt.STOP)
-    open fun smartUpdateSTOP(): Boolean = smartUpdateState(StateInt.STOP)
+    open fun setSTOP(): Boolean = setValue(StateInt.STOP)
+    open fun postSTOP(): Boolean = postValue(StateInt.STOP)
+    open fun smartUpdateSTOP(): Boolean = smartUpdateValue(StateInt.STOP)
 
     // CANCEL
     open fun isCANCELState(): Boolean = isEqual(StateInt.CANCEL)
-    open fun setCANCEL(): Boolean = setState(StateInt.CANCEL)
-    open fun postCANCEL(): Boolean = postState(StateInt.CANCEL)
-    open fun smartUpdateCANCEL(): Boolean = smartUpdateState(StateInt.CANCEL)
+    open fun setCANCEL(): Boolean = setValue(StateInt.CANCEL)
+    open fun postCANCEL(): Boolean = postValue(StateInt.CANCEL)
+    open fun smartUpdateCANCEL(): Boolean = smartUpdateValue(StateInt.CANCEL)
 
     // CREATE
     open fun isCREATEState(): Boolean = isEqual(StateInt.CREATE)
-    open fun setCREATE(): Boolean = setState(StateInt.CREATE)
-    open fun postCREATE(): Boolean = postState(StateInt.CREATE)
-    open fun smartUpdateCREATE(): Boolean = smartUpdateState(StateInt.CREATE)
+    open fun setCREATE(): Boolean = setValue(StateInt.CREATE)
+    open fun postCREATE(): Boolean = postValue(StateInt.CREATE)
+    open fun smartUpdateCREATE(): Boolean = smartUpdateValue(StateInt.CREATE)
 
     // DESTROY
     open fun isDESTROYState(): Boolean = isEqual(StateInt.DESTROY)
-    open fun setDESTROY(): Boolean = setState(StateInt.DESTROY)
-    open fun postDESTROY(): Boolean = postState(StateInt.DESTROY)
-    open fun smartUpdateDESTROY(): Boolean = smartUpdateState(StateInt.DESTROY)
+    open fun setDESTROY(): Boolean = setValue(StateInt.DESTROY)
+    open fun postDESTROY(): Boolean = postValue(StateInt.DESTROY)
+    open fun smartUpdateDESTROY(): Boolean = smartUpdateValue(StateInt.DESTROY)
 
     // RECYCLE
     open fun isRECYCLEState(): Boolean = isEqual(StateInt.RECYCLE)
-    open fun setRECYCLE(): Boolean = setState(StateInt.RECYCLE)
-    open fun postRECYCLE(): Boolean = postState(StateInt.RECYCLE)
-    open fun smartUpdateRECYCLE(): Boolean = smartUpdateState(StateInt.RECYCLE)
+    open fun setRECYCLE(): Boolean = setValue(StateInt.RECYCLE)
+    open fun postRECYCLE(): Boolean = postValue(StateInt.RECYCLE)
+    open fun smartUpdateRECYCLE(): Boolean = smartUpdateValue(StateInt.RECYCLE)
 
     // INIT
     open fun isINITState(): Boolean = isEqual(StateInt.INIT)
-    open fun setINIT(): Boolean = setState(StateInt.INIT)
-    open fun postINIT(): Boolean = postState(StateInt.INIT)
-    open fun smartUpdateINIT(): Boolean = smartUpdateState(StateInt.INIT)
+    open fun setINIT(): Boolean = setValue(StateInt.INIT)
+    open fun postINIT(): Boolean = postValue(StateInt.INIT)
+    open fun smartUpdateINIT(): Boolean = smartUpdateValue(StateInt.INIT)
 
     // ENABLED
     open fun isENABLEDState(): Boolean = isEqual(StateInt.ENABLED)
-    open fun setENABLED(): Boolean = setState(StateInt.ENABLED)
-    open fun postENABLED(): Boolean = postState(StateInt.ENABLED)
-    open fun smartUpdateENABLED(): Boolean = smartUpdateState(StateInt.ENABLED)
+    open fun setENABLED(): Boolean = setValue(StateInt.ENABLED)
+    open fun postENABLED(): Boolean = postValue(StateInt.ENABLED)
+    open fun smartUpdateENABLED(): Boolean = smartUpdateValue(StateInt.ENABLED)
 
     // ENABLING
     open fun isENABLINGState(): Boolean = isEqual(StateInt.ENABLING)
-    open fun setENABLING(): Boolean = setState(StateInt.ENABLING)
-    open fun postENABLING(): Boolean = postState(StateInt.ENABLING)
-    open fun smartUpdateENABLING(): Boolean = smartUpdateState(StateInt.ENABLING)
+    open fun setENABLING(): Boolean = setValue(StateInt.ENABLING)
+    open fun postENABLING(): Boolean = postValue(StateInt.ENABLING)
+    open fun smartUpdateENABLING(): Boolean = smartUpdateValue(StateInt.ENABLING)
 
     // DISABLED
     open fun isDISABLEDState(): Boolean = isEqual(StateInt.DISABLED)
-    open fun setDISABLED(): Boolean = setState(StateInt.DISABLED)
-    open fun postDISABLED(): Boolean = postState(StateInt.DISABLED)
-    open fun smartUpdateDISABLED(): Boolean = smartUpdateState(StateInt.DISABLED)
+    open fun setDISABLED(): Boolean = setValue(StateInt.DISABLED)
+    open fun postDISABLED(): Boolean = postValue(StateInt.DISABLED)
+    open fun smartUpdateDISABLED(): Boolean = smartUpdateValue(StateInt.DISABLED)
 
     // DISABLING
     open fun isDISABLINGState(): Boolean = isEqual(StateInt.DISABLING)
-    open fun setDISABLING(): Boolean = setState(StateInt.DISABLING)
-    open fun postDISABLING(): Boolean = postState(StateInt.DISABLING)
-    open fun smartUpdateDISABLING(): Boolean = smartUpdateState(StateInt.DISABLING)
+    open fun setDISABLING(): Boolean = setValue(StateInt.DISABLING)
+    open fun postDISABLING(): Boolean = postValue(StateInt.DISABLING)
+    open fun smartUpdateDISABLING(): Boolean = smartUpdateValue(StateInt.DISABLING)
 
     // CONNECTED
     open fun isCONNECTEDState(): Boolean = isEqual(StateInt.CONNECTED)
-    open fun setCONNECTED(): Boolean = setState(StateInt.CONNECTED)
-    open fun postCONNECTED(): Boolean = postState(StateInt.CONNECTED)
-    open fun smartUpdateCONNECTED(): Boolean = smartUpdateState(StateInt.CONNECTED)
+    open fun setCONNECTED(): Boolean = setValue(StateInt.CONNECTED)
+    open fun postCONNECTED(): Boolean = postValue(StateInt.CONNECTED)
+    open fun smartUpdateCONNECTED(): Boolean = smartUpdateValue(StateInt.CONNECTED)
 
     // CONNECTING
     open fun isCONNECTINGState(): Boolean = isEqual(StateInt.CONNECTING)
-    open fun setCONNECTING(): Boolean = setState(StateInt.CONNECTING)
-    open fun postCONNECTING(): Boolean = postState(StateInt.CONNECTING)
-    open fun smartUpdateCONNECTING(): Boolean = smartUpdateState(StateInt.CONNECTING)
+    open fun setCONNECTING(): Boolean = setValue(StateInt.CONNECTING)
+    open fun postCONNECTING(): Boolean = postValue(StateInt.CONNECTING)
+    open fun smartUpdateCONNECTING(): Boolean = smartUpdateValue(StateInt.CONNECTING)
 
     // DISCONNECTED
     open fun isDISCONNECTEDState(): Boolean = isEqual(StateInt.DISCONNECTED)
-    open fun setDISCONNECTED(): Boolean = setState(StateInt.DISCONNECTED)
-    open fun postDISCONNECTED(): Boolean = postState(StateInt.DISCONNECTED)
-    open fun smartUpdateDISCONNECTED(): Boolean = smartUpdateState(StateInt.DISCONNECTED)
+    open fun setDISCONNECTED(): Boolean = setValue(StateInt.DISCONNECTED)
+    open fun postDISCONNECTED(): Boolean = postValue(StateInt.DISCONNECTED)
+    open fun smartUpdateDISCONNECTED(): Boolean = smartUpdateValue(StateInt.DISCONNECTED)
 
     // SUSPENDED
     open fun isSUSPENDEDState(): Boolean = isEqual(StateInt.SUSPENDED)
-    open fun setSUSPENDED(): Boolean = setState(StateInt.SUSPENDED)
-    open fun postSUSPENDED(): Boolean = postState(StateInt.SUSPENDED)
-    open fun smartUpdateSUSPENDED(): Boolean = smartUpdateState(StateInt.SUSPENDED)
+    open fun setSUSPENDED(): Boolean = setValue(StateInt.SUSPENDED)
+    open fun postSUSPENDED(): Boolean = postValue(StateInt.SUSPENDED)
+    open fun smartUpdateSUSPENDED(): Boolean = smartUpdateValue(StateInt.SUSPENDED)
 
     // UNKNOWN
     open fun isUNKNOWNState(): Boolean = isEqual(StateInt.UNKNOWN)
-    open fun setUNKNOWN(): Boolean = setState(StateInt.UNKNOWN)
-    open fun postUNKNOWN(): Boolean = postState(StateInt.UNKNOWN)
-    open fun smartUpdateUNKNOWN(): Boolean = smartUpdateState(StateInt.UNKNOWN)
+    open fun setUNKNOWN(): Boolean = setValue(StateInt.UNKNOWN)
+    open fun postUNKNOWN(): Boolean = postValue(StateInt.UNKNOWN)
+    open fun smartUpdateUNKNOWN(): Boolean = smartUpdateValue(StateInt.UNKNOWN)
 
     // INSERT
     open fun isINSERTState(): Boolean = isEqual(StateInt.INSERT)
-    open fun setINSERT(): Boolean = setState(StateInt.INSERT)
-    open fun postINSERT(): Boolean = postState(StateInt.INSERT)
-    open fun smartUpdateINSERT(): Boolean = smartUpdateState(StateInt.INSERT)
+    open fun setINSERT(): Boolean = setValue(StateInt.INSERT)
+    open fun postINSERT(): Boolean = postValue(StateInt.INSERT)
+    open fun smartUpdateINSERT(): Boolean = smartUpdateValue(StateInt.INSERT)
 
     // DELETE
     open fun isDELETEState(): Boolean = isEqual(StateInt.DELETE)
-    open fun setDELETE(): Boolean = setState(StateInt.DELETE)
-    open fun postDELETE(): Boolean = postState(StateInt.DELETE)
-    open fun smartUpdateDELETE(): Boolean = smartUpdateState(StateInt.DELETE)
+    open fun setDELETE(): Boolean = setValue(StateInt.DELETE)
+    open fun postDELETE(): Boolean = postValue(StateInt.DELETE)
+    open fun smartUpdateDELETE(): Boolean = smartUpdateValue(StateInt.DELETE)
 
     // UPDATE
     open fun isUPDATEState(): Boolean = isEqual(StateInt.UPDATE)
-    open fun setUPDATE(): Boolean = setState(StateInt.UPDATE)
-    open fun postUPDATE(): Boolean = postState(StateInt.UPDATE)
-    open fun smartUpdateUPDATE(): Boolean = smartUpdateState(StateInt.UPDATE)
+    open fun setUPDATE(): Boolean = setValue(StateInt.UPDATE)
+    open fun postUPDATE(): Boolean = postValue(StateInt.UPDATE)
+    open fun smartUpdateUPDATE(): Boolean = smartUpdateValue(StateInt.UPDATE)
 
     // SELECT
     open fun isSELECTState(): Boolean = isEqual(StateInt.SELECT)
-    open fun setSELECT(): Boolean = setState(StateInt.SELECT)
-    open fun postSELECT(): Boolean = postState(StateInt.SELECT)
-    open fun smartUpdateSELECT(): Boolean = smartUpdateState(StateInt.SELECT)
+    open fun setSELECT(): Boolean = setValue(StateInt.SELECT)
+    open fun postSELECT(): Boolean = postValue(StateInt.SELECT)
+    open fun smartUpdateSELECT(): Boolean = smartUpdateValue(StateInt.SELECT)
 
     // ENCRYPT
     open fun isENCRYPTState(): Boolean = isEqual(StateInt.ENCRYPT)
-    open fun setENCRYPT(): Boolean = setState(StateInt.ENCRYPT)
-    open fun postENCRYPT(): Boolean = postState(StateInt.ENCRYPT)
-    open fun smartUpdateENCRYPT(): Boolean = smartUpdateState(StateInt.ENCRYPT)
+    open fun setENCRYPT(): Boolean = setValue(StateInt.ENCRYPT)
+    open fun postENCRYPT(): Boolean = postValue(StateInt.ENCRYPT)
+    open fun smartUpdateENCRYPT(): Boolean = smartUpdateValue(StateInt.ENCRYPT)
 
     // DECRYPT
     open fun isDECRYPTState(): Boolean = isEqual(StateInt.DECRYPT)
-    open fun setDECRYPT(): Boolean = setState(StateInt.DECRYPT)
-    open fun postDECRYPT(): Boolean = postState(StateInt.DECRYPT)
-    open fun smartUpdateDECRYPT(): Boolean = smartUpdateState(StateInt.DECRYPT)
+    open fun setDECRYPT(): Boolean = setValue(StateInt.DECRYPT)
+    open fun postDECRYPT(): Boolean = postValue(StateInt.DECRYPT)
+    open fun smartUpdateDECRYPT(): Boolean = smartUpdateValue(StateInt.DECRYPT)
 
     // RESET
     open fun isRESETState(): Boolean = isEqual(StateInt.RESET)
-    open fun setRESET(): Boolean = setState(StateInt.RESET)
-    open fun postRESET(): Boolean = postState(StateInt.RESET)
-    open fun smartUpdateRESET(): Boolean = smartUpdateState(StateInt.RESET)
+    open fun setRESET(): Boolean = setValue(StateInt.RESET)
+    open fun postRESET(): Boolean = postValue(StateInt.RESET)
+    open fun smartUpdateRESET(): Boolean = smartUpdateValue(StateInt.RESET)
 
     // CLOSE
     open fun isCLOSEState(): Boolean = isEqual(StateInt.CLOSE)
-    open fun setCLOSE(): Boolean = setState(StateInt.CLOSE)
-    open fun postCLOSE(): Boolean = postState(StateInt.CLOSE)
-    open fun smartUpdateCLOSE(): Boolean = smartUpdateState(StateInt.CLOSE)
+    open fun setCLOSE(): Boolean = setValue(StateInt.CLOSE)
+    open fun postCLOSE(): Boolean = postValue(StateInt.CLOSE)
+    open fun smartUpdateCLOSE(): Boolean = smartUpdateValue(StateInt.CLOSE)
 
     // OPEN
     open fun isOPENState(): Boolean = isEqual(StateInt.OPEN)
-    open fun setOPEN(): Boolean = setState(StateInt.OPEN)
-    open fun postOPEN(): Boolean = postState(StateInt.OPEN)
-    open fun smartUpdateOPEN(): Boolean = smartUpdateState(StateInt.OPEN)
+    open fun setOPEN(): Boolean = setValue(StateInt.OPEN)
+    open fun postOPEN(): Boolean = postValue(StateInt.OPEN)
+    open fun smartUpdateOPEN(): Boolean = smartUpdateValue(StateInt.OPEN)
 
     // EXIT
     open fun isEXITState(): Boolean = isEqual(StateInt.EXIT)
-    open fun setEXIT(): Boolean = setState(StateInt.EXIT)
-    open fun postEXIT(): Boolean = postState(StateInt.EXIT)
-    open fun smartUpdateEXIT(): Boolean = smartUpdateState(StateInt.EXIT)
+    open fun setEXIT(): Boolean = setValue(StateInt.EXIT)
+    open fun postEXIT(): Boolean = postValue(StateInt.EXIT)
+    open fun smartUpdateEXIT(): Boolean = smartUpdateValue(StateInt.EXIT)
 
     // NEXT
     open fun isNEXTState(): Boolean = isEqual(StateInt.NEXT)
-    open fun setNEXT(): Boolean = setState(StateInt.NEXT)
-    open fun postNEXT(): Boolean = postState(StateInt.NEXT)
-    open fun smartUpdateNEXT(): Boolean = smartUpdateState(StateInt.NEXT)
+    open fun setNEXT(): Boolean = setValue(StateInt.NEXT)
+    open fun postNEXT(): Boolean = postValue(StateInt.NEXT)
+    open fun smartUpdateNEXT(): Boolean = smartUpdateValue(StateInt.NEXT)
 
     // NONE
     open fun isNONEState(): Boolean = isEqual(StateInt.NONE)
-    open fun setNONE(): Boolean = setState(StateInt.NONE)
-    open fun postNONE(): Boolean = postState(StateInt.NONE)
-    open fun smartUpdateNONE(): Boolean = smartUpdateState(StateInt.NONE)
+    open fun setNONE(): Boolean = setValue(StateInt.NONE)
+    open fun postNONE(): Boolean = postValue(StateInt.NONE)
+    open fun smartUpdateNONE(): Boolean = smartUpdateValue(StateInt.NONE)
 
     // FINISH
     open fun isFINISHState(): Boolean = isEqual(StateInt.FINISH)
-    open fun setFINISH(): Boolean = setState(StateInt.FINISH)
-    open fun postFINISH(): Boolean = postState(StateInt.FINISH)
-    open fun smartUpdateFINISH(): Boolean = smartUpdateState(StateInt.FINISH)
+    open fun setFINISH(): Boolean = setValue(StateInt.FINISH)
+    open fun postFINISH(): Boolean = postValue(StateInt.FINISH)
+    open fun smartUpdateFINISH(): Boolean = smartUpdateValue(StateInt.FINISH)
 
     // WAITING
     open fun isWAITINGState(): Boolean = isEqual(StateInt.WAITING)
-    open fun setWAITING(): Boolean = setState(StateInt.WAITING)
-    open fun postWAITING(): Boolean = postState(StateInt.WAITING)
-    open fun smartUpdateWAITING(): Boolean = smartUpdateState(StateInt.WAITING)
+    open fun setWAITING(): Boolean = setValue(StateInt.WAITING)
+    open fun postWAITING(): Boolean = postValue(StateInt.WAITING)
+    open fun smartUpdateWAITING(): Boolean = smartUpdateValue(StateInt.WAITING)
 
     // COMPLETE
     open fun isCOMPLETEState(): Boolean = isEqual(StateInt.COMPLETE)
-    open fun setCOMPLETE(): Boolean = setState(StateInt.COMPLETE)
-    open fun postCOMPLETE(): Boolean = postState(StateInt.COMPLETE)
-    open fun smartUpdateCOMPLETE(): Boolean = smartUpdateState(StateInt.COMPLETE)
+    open fun setCOMPLETE(): Boolean = setValue(StateInt.COMPLETE)
+    open fun postCOMPLETE(): Boolean = postValue(StateInt.COMPLETE)
+    open fun smartUpdateCOMPLETE(): Boolean = smartUpdateValue(StateInt.COMPLETE)
 }