+++ lib/DevSimple/CHANGELOG.md
@@ -6,7 +6,7 @@ Version 1.0.5 *(2026-01-19)*
 
 * `[Add]` 新增 AdapterModel 通用模型
 
-* `[Add]` 新增 StateLiveData、StateIntLiveData
+* `[Add]` 新增 ValueLiveData、StateIntLiveData
 
 Version 1.0.4 *(2025-10-12)*
 ----------------------------